fix: don't print that lockfile is up to date when it is not
close #6544
