+++ .changeset/curly-starfishes-end.md
@@ -1,0 +1,7 @@
+---
+"@pnpm/headless": major
+"@pnpm/core": patch
+"pnpm": patch
+---
+
+Don't print "Lockfile is up-to-date" message before finishing all the lockfile checks [#6544](https://github.com/pnpm/pnpm/issues/6544).
